Move constants to core-ethereum/constants
