--- packages/core-ethereum/src/indexer/constants.ts
@@ -1,1 +1,0 @@
-export const INDEXER_TIMEOUT = 900000 // 15 minutes
